refactor(wallet-sdk): move session-keys into lib
